# This is a combination of two commits. # The first commit's message is:
Removed dummy return error for base lazy loader, fixes issue #92, thanks kzimir for noticing this.


# This is the 2nd commit message:

Implemented connection pooling per hostname.

Up to now, Bulkloader would not take advantage of the fact that when making requests against multiple hosts,
both the client and the server would limit the number (usually) to 2 connections. So if you were loading 8
items from 4 hosts, you'd have the same performance as loading 8 items from the same host.

Now BulkLoader is smart enough to load 2 items (as the maxConnenctionsPerHost property defaults to) per
hostname and still repect the numConnections you'd like to open at most (idenpendent of hostnames).

This should give better loading times, as more requests are sent to the servers earlier on. This should also
improve the wait time for bytesTotal and thus progress indication, as the browser will sent the response headers
earlier on.

Of course, this is only useful if you are making requests against different hostnames. If not, performance should be
unaltered.


git-svn-id: http://bulk-loader.googlecode.com/svn/trunk@272 a142b2d9-a43a-0410-abc3-b9f053a35bbf