Added the BulkLoader visual inspector.
This is a visual FLEX application that can hook up with bulkloader intances and
provide valuable info on them.

If you are using BulkLoader in a flex application , just intantiate the BulkLoaderInspectorPanel class, and you're all set.
If you are using BulkLoader inside a regular, non flex swf, then you should load the BulkLoaderInspectorPanel.swf . This can be done using the Bootstrapper.as class, which will manage to load a flex swf into your hosting regular swf.

git-svn-id: http://bulk-loader.googlecode.com/svn/trunk@290 a142b2d9-a43a-0410-abc3-b9f053a35bbf